+++ fuzz/fuzz_targets/fuzz_dateparse.rs
@@ -1,0 +1,10 @@
+use bolero::check;
+
+fn main() {
+    check!().for_each(|data: &[u8]| match std::str::from_utf8(data) {
+        Ok(input) => {
+            mailparse::dateparse(input).ok();
+        }
+        Err(_err) => {}
+    });
+}