+++ fuzz/fuzz_targets/fuzz_mailparse.rs
@@ -1,0 +1,7 @@
+use bolero::check;
+
+fn main() {
+    check!().for_each(|data: &[u8]| {
+        mailparse::parse_mail(data).ok();
+    });
+}