+++ src/location.rs
@@ -1,7 +1,5 @@
 //! Location handling.
 
-#![allow(missing_docs)]
-
 use std::convert::TryFrom;
 use std::time::Duration;
 
@@ -20,32 +18,63 @@ use crate::mimeparser::SystemMessage;
 use crate::stock_str;
 use crate::tools::{duration_to_str, time};
 
-/// Location record
+/// Location record.
 #[derive(Debug, Clone, Default)]
 pub struct Location {
+    /// Row ID of the location.
     pub location_id: u32,
+
+    /// Location latitude.
     pub latitude: f64,
+
+    /// Location longitude.
     pub longitude: f64,
+
+    /// Nonstandard `accuracy` attribute of the `coordinates` tag.
     pub accuracy: f64,
+
+    /// Location timestamp in seconds.
     pub timestamp: i64,
+
+    /// Contact ID.
     pub contact_id: ContactId,
+
+    /// Message ID.
     pub msg_id: u32,
+
+    /// Chat ID.
     pub chat_id: ChatId,
+
+    /// A marker string, such as an emoji, to be displayed on top of the location.
     pub marker: Option<String>,
+
+    /// Whether location is independent, i.e. not part of the path.
     pub independent: u32,
 }
 
 impl Location {
+    /// Creates a new empty location.
     pub fn new() -> Self {
         Default::default()
     }
 }
 
+/// KML document.
+///
+/// See <https://www.ogc.org/standards/kml/> for the standard and
+/// <https://developers.google.com/kml> for documentation.
 #[derive(Debug, Clone, Default)]
 pub struct Kml {
+    /// Nonstandard `addr` attribute of the `Document` tag storing the user email address.
     pub addr: Option<String>,
+
+    /// Placemarks.
     pub locations: Vec<Location>,
+
+    /// Currently parsed XML tag.
     tag: KmlTag,
+
+    /// Currently parsed placemark.
     pub curr: Location,
 }
 
@@ -62,10 +91,12 @@ bitflags! {
 }
 
 impl Kml {
+    /// Creates a new empty KML document.
     pub fn new() -> Self {
         Default::default()
     }
 
+    /// Parses a KML document.
     pub fn parse(to_parse: &[u8]) -> Result<Self> {
         ensure!(to_parse.len() <= 1024 * 1024, "kml-file is too large");
 
@@ -78,7 +109,7 @@ impl Kml {
         let mut buf = Vec::new();
 
         loop {
-            match reader.read_event(&mut buf).with_context(|| {
+            match reader.read_event_into(&mut buf).with_context(|| {
                 format!(
                     "location parsing error at position {}",
                     reader.buffer_position()
@@ -86,7 +117,7 @@ impl Kml {
             })? {
                 quick_xml::events::Event::Start(ref e) => kml.starttag_cb(e, &reader),
                 quick_xml::events::Event::End(ref e) => kml.endtag_cb(e),
-                quick_xml::events::Event::Text(ref e) => kml.text_cb(e, &reader),
+                quick_xml::events::Event::Text(ref e) => kml.text_cb(e),
                 quick_xml::events::Event::Eof => break,
                 _ => (),
             }
@@ -96,9 +127,9 @@ impl Kml {
         Ok(kml)
     }
 
-    fn text_cb<B: std::io::BufRead>(&mut self, event: &BytesText, reader: &quick_xml::Reader<B>) {
+    fn text_cb(&mut self, event: &BytesText) {
         if self.tag.contains(KmlTag::WHEN) || self.tag.contains(KmlTag::COORDINATES) {
-            let val = event.unescape_and_decode(reader).unwrap_or_default();
+            let val = event.unescape().unwrap_or_default();
 
             let val = val.replace(['\n', '\r', '\t', ' '], "");
 
@@ -127,7 +158,9 @@ impl Kml {
     }
 
     fn endtag_cb(&mut self, event: &BytesEnd) {
-        let tag = String::from_utf8_lossy(event.name()).trim().to_lowercase();
+        let tag = String::from_utf8_lossy(event.name().as_ref())
+            .trim()
+            .to_lowercase();
 
         if tag == "placemark" {
             if self.tag.contains(KmlTag::PLACEMARK)
@@ -147,14 +180,20 @@ impl Kml {
         event: &BytesStart,
         reader: &quick_xml::Reader<B>,
     ) {
-        let tag = String::from_utf8_lossy(event.name()).trim().to_lowercase();
+        let tag = String::from_utf8_lossy(event.name().as_ref())
+            .trim()
+            .to_lowercase();
         if tag == "document" {
-            if let Some(addr) = event
-                .attributes()
-                .filter_map(|a| a.ok())
-                .find(|attr| String::from_utf8_lossy(attr.key).trim().to_lowercase() == "addr")
-            {
-                self.addr = addr.unescape_and_decode_value(reader).ok();
+            if let Some(addr) = event.attributes().filter_map(|a| a.ok()).find(|attr| {
+                String::from_utf8_lossy(attr.key.as_ref())
+                    .trim()
+                    .to_lowercase()
+                    == "addr"
+            }) {
+                self.addr = addr
+                    .decode_and_unescape_value(reader)
+                    .ok()
+                    .map(|a| a.into_owned());
             }
         } else if tag == "placemark" {
             self.tag = KmlTag::PLACEMARK;
@@ -172,12 +211,17 @@ impl Kml {
             self.tag = KmlTag::PLACEMARK | KmlTag::POINT | KmlTag::COORDINATES;
             if let Some(acc) = event.attributes().find(|attr| {
                 attr.as_ref()
-                    .map(|a| String::from_utf8_lossy(a.key).trim().to_lowercase() == "accuracy")
+                    .map(|a| {
+                        String::from_utf8_lossy(a.key.as_ref())
+                            .trim()
+                            .to_lowercase()
+                            == "accuracy"
+                    })
                     .unwrap_or_default()
             }) {
                 let v = acc
                     .unwrap()
-                    .unescape_and_decode_value(reader)
+                    .decode_and_unescape_value(reader)
                     .unwrap_or_default();
 
                 self.curr.accuracy = v.trim().parse().unwrap_or_default();
@@ -259,6 +303,7 @@ pub async fn is_sending_locations_to_chat(
     Ok(exists)
 }
 
+/// Sets current location of the user device.
 pub async fn set(context: &Context, latitude: f64, longitude: f64, accuracy: f64) -> bool {
     if latitude == 0.0 && longitude == 0.0 {
         return true;
@@ -306,6 +351,7 @@ pub async fn set(context: &Context, latitude: f64, longitude: f64, accuracy: f64
     continue_streaming
 }
 
+/// Searches for locations in the given time range, optionally filtering by chat and contact IDs.
 pub async fn get_range(
     context: &Context,
     chat_id: Option<ChatId>,
@@ -396,6 +442,7 @@ pub async fn delete_all(context: &Context) -> Result<()> {
     Ok(())
 }
 
+/// Returns `location.kml` contents.
 pub async fn get_kml(context: &Context, chat_id: ChatId) -> Result<(String, u32)> {
     let mut last_added_location_id = 0;
 
@@ -481,6 +528,7 @@ fn get_kml_timestamp(utc: i64) -> String {
         .to_string()
 }
 
+/// Returns a KML document containing a single location with the given timestamp and coordinates.
 pub fn get_message_kml(timestamp: i64, latitude: f64, longitude: f64) -> String {
     format!(
         "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n\
@@ -498,6 +546,7 @@ pub fn get_message_kml(timestamp: i64, latitude: f64, longitude: f64) -> String 
     )
 }
 
+/// Sets the timestamp of the last time location was sent in the chat.
 pub async fn set_kml_sent_timestamp(
     context: &Context,
     chat_id: ChatId,
@@ -513,6 +562,7 @@ pub async fn set_kml_sent_timestamp(
     Ok(())
 }
 
+/// Sets the location of the message.
 pub async fn set_msg_location_id(context: &Context, msg_id: MsgId, location_id: u32) -> Result<()> {
     context
         .sql