+++ src/summary.rs
@@ -1,7 +1,5 @@
 //! # Message summary for chatlist.
 
-#![allow(missing_docs)]
-
 use crate::chat::Chat;
 use crate::constants::Chattype;
 use crate::contact::{Contact, ContactId};
@@ -54,6 +52,8 @@ pub struct Summary {
 }
 
 impl Summary {
+    /// Constucts chatlist summary
+    /// from the provided message, chat and message author contact snapshots.
     pub async fn new(
         context: &Context,
         msg: &Message,